# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "蓝博客户端api接口文档",
        "description": "API接口返回描述：code=200正常，code=401登录信息错误，code=0其他错误。msg为返回描述，data为返回结构体",
        "contact": {
            "email": "834574377@qq.com"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://192.168.60.42/ptweb/",
            "description": "人人兼职本机测试"
        },
        {
            "url": "https://juliangtec.com/app",
            "description": "邮箱发送服务器"
        },
        {
            "url": "https://api.lanbodsp.com",
            "description": "DSP线上"
        }
    ],
    "paths": {
...
        "/orderlist": {
            "post": {
                "tags": [
                    "订单模块"
                ],
                "summary": "订单列表",
                "operationId": "7fc0b48b24bbc19b8eb0b841555b25fe",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "状态",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_time",
                        "in": "query",
                        "description": "开始时间",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "end_time",
                        "in": "query",
                        "description": "结束时间",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "页码",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{code: int, msg:string, data:[]}"
                    }
                }
            }
        },
        "/getorder": {
            "post": {
                "tags": [
                    "订单模块"
                ],
                "summary": "订单详情",
                "operationId": "d16060571ed9bf891207cc03598869ff",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "编号",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{code: int, msg:string, data:[]}"
                    }
                }
            }
        },
        "/create_order": {
            "post": {
                "tags": [
                    "订单模块"
                ],
                "summary": "报名",
                "operationId": "1d72dcd8aea4b395cfac2ae5bcb5cf28",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "编号",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{code: int, msg:string, data:[]}"
                    }
                }
            }
        },
        "/collect": {
            "post": {
                "tags": [
                    "订单模块"
                ],
                "summary": "收藏",
                "operationId": "18cc57f98aef1e5652b7003ee0b0f71f",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "编号",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{code: int, msg:string, data:[]}"
                    }
                }
            }
        },
        "/collectlist": {
            "post": {
                "tags": [
                    "订单模块"
                ],
                "summary": "收藏列表",
                "operationId": "96d9cbe7fa9c2e42a23f06a560e4f30e",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_time",
                        "in": "query",
                        "description": "开始时间",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "end_time",
                        "in": "query",
                        "description": "结束时间",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "页码",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{code: int, msg:string, data:[]}"
                    }
                }
            }
        },
...
    }
}